Using the youngest child pid to create core file for looping job. Moved get_subprocesses() to psutils module
