+++ pilot/util/loopingjob.py
@@ -14,7 +14,7 @@ from pilot.util.container import execute  #, execute_command
 from pilot.util.filehandling import remove_files, find_latest_modified_file, verify_file_list, copy, list_mod_files
 from pilot.util.parameters import convert_to_int
 from pilot.util.processes import kill_process, find_zombies, handle_zombies, reap_zombies
-from pilot.util.psutils import get_child_processes
+from pilot.util.psutils import get_child_processes, get_subprocesses
 from pilot.util.timing import time_stamp
 
 import os
@@ -91,7 +91,15 @@ def create_core_dump(job):
         logger.warning('cannot create core file since pid or workdir is unknown')
         return
 
-    cmd = f'gdb --pid {job.pid} -ex \'generate-core-file\' -ex quit'
+    # get the pid of the youngest child belonging to the payload
+    pids = get_subprocesses(job.pid)
+    if not pids:
+        pid = job.pid
+        logger.info(f'the payload process ({pid}) has no children')
+    else:
+        logger.info(f'the payload process ({job.pid}) has the following children: {pids}')
+        pid = pids[-1]
+    cmd = f'gdb --pid {pid} -ex \'generate-core-file\' -ex quit'
     exit_code, stdout, stderr = execute(cmd)
     #exit_code = execute_command(cmd)
     if exit_code == 0:
